Extract efficient conversion of $set to jsonb_set
This has been the trickiest thing so far: trying to produce a SQL
expression using jsonb_set that has the exact same effect as MongoDB's
$set operation.

As discussed previously, the key difference is that $set allows the user
to set nested fields even they don't already exist, e.g. 'a.b.c' => 1
will create 'a' and 'b' as objects/hashes automatically. jsonb_set on
the other hand will be a no-op if 'a' and 'b' don't already exist as
objects/hashes.

The previous approach worked by introducing extra jsonb_set calls that
attempt to set any intermediate keys as JSONB hashes _if they don't
already exist_. This is achieved by using the SQL COALESCE function to
set any intermediate keys either to themselves or to an empty hash. The
tricky bit here is that we have to make sure that their value is as it
would be _after_ jsonb_set, e.g. when checking the value of 'b' in an
update to 'a.b.c' we need to have already set 'a' to an empty hash.

As this introduces quite a few calls to jsonb_set, we want to keep this
as efficient as possible by only setting intermediate keys to empty
hashes once even if multiple updates share the same paths, e.g.
'a.b.c.d' => 1, 'a.b.c.e' => 2 should only build 'a', 'b' and 'c' once.

To do this, we explicitly extract a tree builder and walker: one to
convert a $set operation into a tree of paths and one to walk this tree,
emitting the necessary SQL equivalent.

e.g.

   $set: { 'a.b.c' => 1, 'a.b.d' => 2, 'a.e' => 3}

becomes

   {
      ['a'] => {
         ['a', 'b'] => {
            ['a', 'b', 'c'] => 1,
            ['a', 'b', 'd'] => 2
         },
         ['a', 'e'] => 3
      }
   }