Refactor the Set building code using polymorphism
Rather than using a vanilla Ruby hash to store the internal tree
structure of a MongoDB set operation, use explicit Intermediate and
Assignment node classes. This way, we can define the SQL version on each
node type rather than having to mess around with conditional in an
inject.

Speaking of inject, I replaced that with more explicit (but less
idiomatic) code when building a Set from a MongoDB operation as it was
rather tricky to understand and debug previously.