Refactor and add configuration file for mappings
Rather than attempting to replicate every single collection, allow the
user to pass a configuration file with a mapping between MongoDB
namespaces and PostgreSQL tables. This means that operations on
collections will be ignored by default unless they are in the
configuration file.

As we now determine the target table for each operation from the
mapping, drop the SQL methods in favour of operating directly on the
tables themselves. This means there is a pressing need for integration
tests that hit real databases.