Refocus library on oplog to SQL transformation
Rather than including code to configure mappings between MongoDB
namespaces and PostgreSQL tables and apply those transformations,
recording progress as we go, concentrate the library on the core
transformation of MongoDB oplog operations to SQL statements.

This means that the library can be used in the core of an application to
tail an oplog but leaves that to the user.

The other two big changes are:

1. Strip null bytes from any JSON before sending to PostgreSQL to avoid
errors;
2. Use PostgreSQL 9.5's INSERT ON CONFLICT to turn insert operations
into upserts (as they should be idempotent).