Make the 'msg' in a noop optional
As this was required before, I didn't want to change to be suprisingly
returning a nil. Anything that someone was doing with this string before
should still work as expected, even if the message is empty now.

I've not been able to track down the schema for the Oplog, but we've run
into a scenario where we have noops which contain the full object in
'o', and do not include 'msg'. This situation could have come about
because of a bug on our side; we upgraded Mongoid for one of our
applications and it/Mongo(?) started producing these odd noop messages.
Our investigation onto how that happened is on going.